initialize attribute currentUser0 and add a null check in tear down to avoid dependence of assert
git-svn-id: http://vcs.taotesting.com/svn/taoqual/branches/WorlflowEngineExt@1421 793250d4-a296-4af0-ab9f-c080ea916244

+++ test/ProcessExecutionServiceTestCase.php
@@ -38,6 +38,8 @@ class ProcessExecutionServiceTestCase extends UnitTestCase{
 	 */
 	protected $currentUser = null;
 	
+	protected $currentUser0 = null;
+	
 	/**
 	 * initialize a test method
 	 */
@@ -75,7 +77,9 @@ class ProcessExecutionServiceTestCase extends UnitTestCase{
 	}
 	
 	public function tearDown() {
-		$this->currentUser0->delete();
+		if (!is_null($this->$currentUser0)){
+			$this->currentUser0->delete();
+		}
     }
 	
 	/**
